requirements.txt
Flask
Flask-CORS
requests
pywebpush
cryptography
Pillow
gunicorn